Add Promise support
Adds a “promise()” method to the streams, which converts the stream into a Promise. The promise is handled in the same way as the streaming API, meaning the results are just returned in an array instead of embedded in an object.

Additionally, Promises are now supported on the create, get, update, and delete APIs for individual models, as well as the batch get items API.